Don't add newline adding GitHub URL to clipboard
Echo adds a newline, which isn't great for the functionality we
require. However, there is a -n flag that can be passed into echo which
bypasses this append.

For some reason though, `echo -n str | pbcopy` adds '-n str' to the
clipboard, rather than not appending a newline to the echoed string.

So to remove the newline we pipe the echo output through tr, removing
any newline characters.